Final Commit

+++ src/components/profile/post/comments/CommentsSection.jsx
@@ -38,6 +38,11 @@ const CommentsSection = ({ blog }) => {
 
     const handleDeleteComment = async (id) => {
         try {
+            const confirmed = window.confirm(
+                "Are you sure you want to delete this comment?"
+            );
+            if (!confirmed) return;
+
             const response = await api.delete(
                 `${import.meta.env.VITE_SERVER_BASE_URL}/blogs/${
                     blog?.id
